% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{build_r_plotly}
\alias{build_r_plotly}
\title{From}
\usage{
build_r_plotly(plotly_figure)
}
\arguments{
\item{plotly_figure}{The result returned from summarize_result()}
}
\value{
The HTML required for nice display of multiple output tables
}
\description{
From
}
