% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{build_r_plots_all}
\alias{build_r_plots_all}
\title{Provide a summarization of all results}
\usage{
build_r_plots_all(results)
}
\arguments{
\item{results}{A list of result objects}
}
\value{
A numeric vector giving number of characters (code points) in each
   element of the character vector. Missing string have missing length.
}
\description{
Provide a summarization of all results
}
