% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{display_report}
\alias{display_report}
\title{Produce RMarkdown-compatible output of all results}
\usage{
display_report(processed_results)
}
\arguments{
\item{processed_results}{A list of processed result objects}
}
\value{
A formatted rmarkdown
}
\description{
Produce RMarkdown-compatible output of all results
}
