% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pretty_print_all_plotly}
\alias{pretty_print_all_plotly}
\title{Provide a pretty printing of all processed results}
\usage{
pretty_print_all_plotly(processed_results)
}
\arguments{
\item{processed_results}{A list of result objects}
}
\value{
A numeric vector giving number of characters (code points) in each
   element of the character vector. Missing string have missing length.
}
\description{
Provide a pretty printing of all processed results
}
