% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pretty_print_tables}
\alias{pretty_print_tables}
\title{Pretty print a list of tables}
\usage{
pretty_print_tables(datasets)
}
\arguments{
\item{datasets}{The result returned from summarize_results()}
}
\value{
The HTML required for nice display of multiple output tables
}
\description{
Pretty print a list of tables
}
