% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{print_summary_tables}
\alias{print_summary_tables}
\title{Print a summary table of the ValidMind results}
\usage{
print_summary_tables(result_summary)
}
\arguments{
\item{result_summary}{A summary of the results}
}
\value{
A data frame containing the summary of the ValidMind results
}
\description{
Print a summary table of the ValidMind results
}
