% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tests.R
\name{run_test}
\alias{run_test}
\title{Run a Test Function with Named Inputs}
\usage{
run_test(func, inputs)
}
\arguments{
\item{func}{A function to be executed. This should be an R function that can be called with the inputs provided.}

\item{inputs}{A named list of inputs to be passed to the function. The names of the list elements should match 
the argument names expected by the function.}
}
\value{
The result of executing the function with the provided inputs. The return value can be of any type depending 
  on the function being executed.
}
\description{
This function allows you to run a specified function with a given set of named inputs. 
The function will be executed with the inputs provided, and the results will be returned.
}
\examples{
# Define a simple function to add two numbers
add_numbers <- function(a, b) {
  return(a + b)
}

# Run the test function with named inputs
result <- run_test(add_numbers, list(a = 5, b = 3))
print(result)  # Should print 8

\dontrun{
# Example with a more complex function
library(ggplot2)
library(caret)

# Define a function to generate a confusion matrix and plot it
confusion_matrix <- function(model, dataset, response_var = "Exited") {
  y_true <- as.factor(dataset[[response_var]])
  y_pred <- predict(model, dataset)
  
  y_pred <- as.factor(ifelse(y_pred <= 0, 0, 1))
  cm <- confusionMatrix(y_pred, y_true)
  
  cm_data <- as.data.frame(cm$table)
  cm_data$Prediction <- factor(cm_data$Prediction, levels = rev(levels(cm_data$Prediction)))
  
  p <- ggplot(cm_data, aes(x = Prediction, y = Reference, fill = Freq)) +
    geom_tile() +
    geom_text(aes(label = Freq), vjust = 1) +
    scale_fill_gradient(low = "white", high = "red") +
    theme_minimal() +
    labs(title = "Confusion Matrix", x = "Predicted", y = "Actual")
  
  print(p)  # print the plot to display it
  
  return(p)  # return the ggplot object
}

# Assuming 'model' is a trained model and 'dataset' is a data frame
# Run the test function with the confusion matrix example
result <- run_test(confusion_matrix, list(
    model = model, dataset = dataset, response_var = "Exited"
))
}

}
