% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{summarize_metric_result}
\alias{summarize_metric_result}
\title{Provide a summarization of a single metric result}
\usage{
summarize_metric_result(result)
}
\arguments{
\item{result}{The ValidMind result object}
}
\description{
Provide a summarization of a single metric result
}
