% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{summarize_result}
\alias{summarize_result}
\title{Provide a summarization of a single result}
\usage{
summarize_result(result)
}
\arguments{
\item{result}{The result object}
}
\description{
Provide a summarization of a single result
}
