% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{vm}
\alias{vm}
\title{Retrieve a validmind (vm) connection object using reticulate}
\usage{
vm(
  api_key,
  api_secret,
  model,
  python_version = Sys.getenv("VALIDMIND_PYTHON", Sys.which("python")),
  api_host = "http://localhost:3000/api/v1/tracking",
  document = NULL
)
}
\arguments{
\item{api_key}{The ValidMind API key}

\item{api_secret}{The ValidMind API secret}

\item{model}{The ValidMind model}

\item{python_version}{The path to the Python binary to use. Defaults to
the VALIDMIND_PYTHON environment variable, or the system Python.}

\item{api_host}{The ValidMind host, defaulting to local}

\item{document}{The document type to associate with this session
(e.g. "documentation", "validation-report"). Defaults to NULL.}
}
\value{
A validmind connection object, obtained from `reticulate`,
which orchestrates the connection to the ValidMind API
}
\description{
Retrieve a validmind (vm) connection object using reticulate
}
\examples{
\dontrun{
vm_r <- vm(
   api_host="https://app.prod.validmind.ai/api/v1/tracking",
   api_key="<your_api_key_here>",
   api_secret="<your_api_secret_here>",
   model="<your_model_id_here>",
   document="documentation"
 )
}

}
