% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tests.R
\name{vm_test}
\alias{vm_test}
\title{Register a Test Function}
\usage{
vm_test(func_or_id)
}
\arguments{
\item{func_or_id}{A function to be registered as a test, or an id of an existing test. The function should be compatible with the `validmind` test decorator.}
}
\value{
A decorator function that wraps the input function for testing purposes.
}
\description{
This function registers a given function as a test using the `validmind` Python library through the `reticulate` package.
}
\examples{
\dontrun{
my_test_function <- function(self, a = 1, b = 2) {
    return(a + b)
}
registered_test <- vm_test(my_test_function)
}
}
