% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{build_r_plotly}
\alias{build_r_plotly}
\title{Build an R Plotly figure from a JSON representation}
\usage{
build_r_plotly(plotly_figure)
}
\arguments{
\item{plotly_figure}{A nested list containing plotly elements}
}
\value{
An R Plotly object derived from the JSON representation
}
\description{
Build an R Plotly figure from a JSON representation
}
