% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{process_result}
\alias{process_result}
\title{Process a set of ValidMind results into parseable data}
\usage{
process_result(results)
}
\arguments{
\item{results}{A list of ValidMind result objects}
}
\value{
A numeric vector giving number of characters (code points) in each
   element of the character vector. Missing string have missing length.
}
\description{
Process a set of ValidMind results into parseable data
}
