% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{save_model}
\alias{save_model}
\title{Save a model to a given file path}
\usage{
save_model(model)
}
\arguments{
\item{model}{The model object}
}
\value{
A numeric vector giving number of characters (code points) in each
   element of the character vector. Missing string have missing length.
}
\description{
Save a model to a given file path
}
\examples{
my_model <- lm(Sepal.Width ~ Sepal.Length, data = iris)
save_model(my_model)
}
