% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{save_model}
\alias{save_model}
\title{Save an R model to a temporary file}
\usage{
save_model(model)
}
\arguments{
\item{model}{The R model object to be saved.}
}
\value{
A string representing the full file path to the saved `.RData` file.
}
\description{
This function saves a given R model object to a randomly named `.RData` file
in the `/tmp/` directory. The file is saved with a unique name generated
using random letters.
}
\examples{
model <- lm(mpg ~ cyl, data = mtcars)
file_path <- save_model(model)

}
