% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{summarize_result}
\alias{summarize_result}
\title{Provide a summarization of a single result (test or metric)}
\usage{
summarize_result(result)
}
\arguments{
\item{result}{The ValidMind result object}
}
\value{
Based on the type of `result`, either A list containing the summary
of the ValidMind results, or a list containing the summary of the ValidMind
results
}
\description{
Provide a summarization of a single result (test or metric)
}
