% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{summarize_test_result}
\alias{summarize_test_result}
\title{Provide a summarization of a single test result}
\usage{
summarize_test_result(result)
}
\arguments{
\item{result}{The ValidMind result object}
}
\value{
A list containing the summary of the ValidMind test results
}
\description{
Provide a summarization of a single test result
}
