% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vm}
\alias{vm}
\title{Retrieve a validmind (vm) object using reticulate}
\usage{
vm(
  api_key,
  api_secret,
  project,
  python_version,
  api_host = "http://localhost:3000/api/v1/tracking"
)
}
\arguments{
\item{api_key}{The ValidMind API key}

\item{api_secret}{The ValidMind API secret}

\item{project}{The ValidMind project}

\item{python_version}{The Python Version to use}

\item{api_host}{The ValidMind host, defaulting to local}
}
\description{
Retrieve a validmind (vm) object using reticulate
}
